Relocate 'Location' class from 'fileops' to 'utils' module.
